requirements.txt
yfinance
streamlit
pandas
plotly
click
prophet
numpy
scikit-learn
boto3
botocore
bs4
langchain
langchain-community
python-dotenv
langchain-aws
streamlit-chat
PyPDF2
pandas-datareader
anthropic
vnstock3
seaborn
transformers